requirements.txt
Flask
python-dotenv
openai
python-telegram-bot
requests
pinecone